Fix python Keyerror
If the database is empty the fetch doesn't return
any row (obviously...). This situation leads to a
TypeError when trying to read inverter data, and a
KeyError when trying to save the optimizer data.

This fix catches that situation.

Better yet - I could test if the database was empty,
and initialize all tables. But that is one for the
roadmap.